Extract CastMember composable from CreditStrip into its own file.
