+++ Samples/Flix/src/main/java/dev/testify/samples/flix/ui/common/composeables/CreditStrip.kt
@@ -54,44 +54,3 @@ fun CreditStrip(
         CastMember(model = it)
     }
 }
-
-@Composable
-fun CastMember(
-    model: CreditPresentationModel,
-    modifier: Modifier = Modifier,
-    viewAction: ViewAction? = null,
-    onPressed: (() -> Unit)? = null
-) {
-    Surface(
-        modifier = modifier
-            .width(100.dp)
-            .height(200.dp)
-            .padding(5.dp),
-        shadowElevation = 1.dp,
-        tonalElevation = 1.dp
-    ) {
-        Column {
-            AsyncImage(
-                model = model.imagePath,
-                contentDescription = viewAction?.describe(),
-                modifier = Modifier
-                    .clickable(
-                        onPressed != null,
-                        role = Role.Button
-                    ) { onPressed?.invoke() },
-            )
-            Text(
-                text = model.name,
-                style = MaterialTheme.typography.labelMedium,
-                maxLines = 2
-            )
-            Text(
-                text = model.characterName,
-                style = MaterialTheme.typography.labelSmall,
-                fontWeight = FontWeight.Light,
-                maxLines = 2,
-                overflow = TextOverflow.Clip
-            )
-        }
-    }
-}
